requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bump-pydantic==0.7.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.105.0
fastapi-jwt-auth==0.5.0
greenlet==3.0.2
h11==0.14.0
idna==3.6
libcst==1.1.0
Mako==1.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mypy-extensions==1.0.0
psycopg2-binary==2.9.9
pydantic==1.10.13
pydantic_core==2.14.5
Pygments==2.17.2
PyJWT==1.7.1
python-dotenv==1.0.0
PyYAML==6.0.1
rich==13.7.0
sniffio==1.3.0
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
uvicorn==0.24.0.post1
Werkzeug==3.0.1
